% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCPN.R
\name{fitCPN}
\alias{fitCPN}
\title{Fit a Bayesian network model to cancer mutation data}
\usage{
fitCPN(
  df,
  model = "CBN",
  algorithm = "DP",
  k = 3,
  epsilon = min(colMeans(df)/2),
  ngen = 100,
  popsize = 100,
  verbose = TRUE
)
}
\arguments{
\item{df}{A dataframe or matrix of binary mutation data. The columns should be 
mutations and the rows should be patients.}

\item{model}{Whether to fit a network with "conjuctive" interactions (CBN), or
"disjunctive" interactions (DBN).}

\item{algorithm}{The algorithm used to infer the network. Default is
an optimal dynamic programming ("DP") algorithm, which is only feasible when the number of mutations is
smaller than 30. For larger datasets, use the genetic algorithm ("GA"), which 
gives an approximate solution.}

\item{k}{The in-degree bound on the estimated network.}

\item{epsilon}{Penalty term (for mutations not conforming to the estimated network).
Default}

\item{ngen}{For genetic algorithm only: number of generations.}

\item{popsize}{For genetic algorithm only: initial population size.}

\item{verbose}{Whether or not to print to the console.}
}
\value{
A list with components 
\itemize{
\item \code{edgelist} - A vector of length 2*k listing
the vertices in the k inferred edges.
\item \code{score} - The maximum likelihood value.
\item \code{graph} - The maximum likelihood graph saved as 
an igraph object.
}
}
\description{
Fit a Bayesian network model to cancer evolution data.
}
\examples{
## Load the example data
data("example")

## Fit a conjunctive Bayesian network (CBN)
out <- fitCPN(example, epsilon=0.01)

## Plot the graph
plotCPN(out)

}
\references{
Nicol, PB, Coombes, KR, Deaver, C, Chkrebtii, O, Paul, S, Toland, AE, Asiaee, A. 
Oncogenetic network estimation with disjunctive Bayesian networks. Comp. Sys. Onco. 2021; 1:e1027.
 https://doi.org/10.1002/cso.21027
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
