% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCPN.R
\name{Lik.genotype}
\alias{Lik.genotype}
\title{Compute probability of a given genotype.}
\usage{
Lik.genotype(fit, genotype)
}
\arguments{
\item{fit}{An object obtained from plotCPN}

\item{genotype}{A binary vector with length equal to the 
number of mutations.}
}
\value{
A scalar value, equal to the probability of the
input genotype under the inferred model parameters.
}
\description{
Compute probability of a given genotype.
}
\examples{
## TODO

}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
