% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCPN.R
\name{plotCPN}
\alias{plotCPN}
\title{Plot a cancer progression network}
\usage{
plotCPN(fit)
}
\arguments{
\item{fit}{An object obtained from plotCPN.}
}
\description{
This function plots a cancer progression network inferred 
in plotCPN.
}
\examples{
## Load the example data
data("example")

## Fit a conjunctive Bayesian network (CBN)
out <- fitCPN(example, epsilon=0.01)

## Plot the graph
plotCPN(out)

}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
