% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCPN.R
\name{fitCPN}
\alias{fitCPN}
\title{Fit a Bayesian network model to cancer mutation data}
\usage{
fitCPN(
  df,
  model = "CBN",
  algorithm = "DP",
  k = 3,
  epsilon = 0.025,
  ngen = 100,
  popsize = 100,
  verbose = TRUE
)
}
\arguments{
\item{df}{Number of cells in the tumor.}

\item{model}{Cell division rate.}

\item{algorithm}{Cell death rate.}

\item{k}{Mutation rate. When a cell divides, both daughter cell acquire \eqn{Pois(u)} genetic alterations}

\item{epsilon}{The probability that a genetic alteration is a driver mutation.}

\item{ngen}{The selective advantage conferred to a driver mutation. A cell with k
driver mutations is given birth rate \eqn{bs^k}.}

\item{popsize}{Edge list for a directed acyclic graph describing possible transitions between states. See
\code{\link{progressionChain}()} for an example of a valid input matrix.}

\item{verbose}{Whether or not to print simulation details to the R console.}
}
\value{
A list with components 
\itemize{
\item \code{edgelist} - A data frame containing the information for the simulated cells. (x,y,z) position, allele ID number
(note that 0 is the wild-type allele),
number of genetic alterations, and Euclidean distance from origin are included. 
\item \code{score} - A data frame consisting of the mutation ID number, the count of the mutation within the population, 
and the mutation allele frequency (which is the count divided by N). 
}
}
\description{
TODO
}
\details{
TODO
}
\examples{
## Load the example data
data("example")

## Fit a conjunctive Bayesian network (CBN)
out <- fitCPN(example, epsilon=0.01)

## Plot the graph
plotCPN(out)

}
\references{
TODO
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
